+++ Platform/OpenCore/OpenCoreMisc.c
@@ -115,6 +115,74 @@ OcToolLoadEntry (
   return EFI_SUCCESS;
 }
 
+CONST CHAR8 *
+OcMiscGetVersionString (
+  VOID
+  )
+{
+  UINT32  Month;
+
+  /**
+    Force the assertions in case we forget about them.
+  **/
+  OC_STATIC_ASSERT (
+    L_STR_LEN (OPEN_CORE_VERSION) == 5,
+    "OPEN_CORE_VERSION must follow X.Y.Z format, where X.Y.Z are single digits."
+    );
+
+  OC_STATIC_ASSERT (
+    L_STR_LEN (OPEN_CORE_TARGET) == 3,
+    "OPEN_CORE_TARGET must XYZ format, where XYZ is build target."
+    );
+
+  STATIC CHAR8 mOpenCoreVersion[] = {
+    /* [2]:[0]    = */ OPEN_CORE_TARGET
+    /* [3]        = */ "-"
+    /* [6]:[4]    = */ "XXX"
+    /* [7]        = */ "-"
+    /* [12]:[8]   = */ "YYYY-"
+    /* [15]:[13]  = */ "MM-"
+    /* [17]:[16]  = */ "DD"
+  };
+
+  STATIC BOOLEAN mOpenCoreVersionReady;
+
+  if (!mOpenCoreVersionReady) {
+
+    mOpenCoreVersion[4] = OPEN_CORE_VERSION[0];
+    mOpenCoreVersion[5] = OPEN_CORE_VERSION[2];
+    mOpenCoreVersion[6] = OPEN_CORE_VERSION[4];
+
+    mOpenCoreVersion[8]  = __DATE__[7];
+    mOpenCoreVersion[9]  = __DATE__[8];
+    mOpenCoreVersion[10] = __DATE__[9];
+    mOpenCoreVersion[11] = __DATE__[10];
+
+    Month =
+      (__DATE__[0] == 'J' && __DATE__[1] == 'a' && __DATE__[2] == 'n') ?  1 :
+      (__DATE__[0] == 'F' && __DATE__[1] == 'e' && __DATE__[2] == 'b') ?  2 :
+      (__DATE__[0] == 'M' && __DATE__[1] == 'a' && __DATE__[2] == 'r') ?  3 :
+      (__DATE__[0] == 'A' && __DATE__[1] == 'p' && __DATE__[2] == 'r') ?  4 :
+      (__DATE__[0] == 'M' && __DATE__[1] == 'a' && __DATE__[2] == 'y') ?  5 :
+      (__DATE__[0] == 'J' && __DATE__[1] == 'u' && __DATE__[2] == 'n') ?  6 :
+      (__DATE__[0] == 'J' && __DATE__[1] == 'u' && __DATE__[2] == 'l') ?  7 :
+      (__DATE__[0] == 'A' && __DATE__[1] == 'u' && __DATE__[2] == 'g') ?  8 :
+      (__DATE__[0] == 'S' && __DATE__[1] == 'e' && __DATE__[2] == 'p') ?  9 :
+      (__DATE__[0] == 'O' && __DATE__[1] == 'c' && __DATE__[2] == 't') ? 10 :
+      (__DATE__[0] == 'N' && __DATE__[1] == 'o' && __DATE__[2] == 'v') ? 11 :
+      (__DATE__[0] == 'D' && __DATE__[1] == 'e' && __DATE__[2] == 'c') ? 12 : 0;
+
+    mOpenCoreVersion[13] = Month < 10 ? '0' : '1';
+    mOpenCoreVersion[14] = '0' + (Month % 10);
+    mOpenCoreVersion[16] = __DATE__[4] >= '0' ? __DATE__[4] : '0';
+    mOpenCoreVersion[17] = __DATE__[5];
+
+    mOpenCoreVersionReady = TRUE;
+  }
+
+  return mOpenCoreVersion;
+}
+
 EFI_STATUS
 OcMiscEarlyInit (
   IN  OC_STORAGE_CONTEXT *Storage,
@@ -457,6 +525,10 @@ OcMiscBoot (
   Context->CustomRead         = OcToolLoadEntry;
   Context->PrivilegeContext   = Privilege;
   Context->RequestPrivilege   = OcShowSimplePasswordRequest;
+
+  if ((Config->Misc.Security.ExposeSensitiveData & OCS_EXPOSE_VERSION_UI) != 0) {
+    Context->TitleSuffix      = OcMiscGetVersionString ();
+  }
 
   if (Config->Misc.Boot.ShowPicker) {
     PickerCommand = Context->PickerCommand = OcPickerShowPicker;